Refactor: make App component separate file from render
